Fix read()/write() prototype
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@820 42af7a65-404d-4744-a932-0658087f49c3

+++ fs/fs_read.c
@@ -53,7 +53,7 @@
  * Global Functions
  ****************************************************************************/
 
-int read(int fd, FAR void *buf, unsigned int nbytes)
+ssize_t read(int fd, FAR void *buf, size_t nbytes)
 {
   FAR struct filelist *list;
   int ret = EBADF;
